Sanitize timer lifetime
In order to stop, the timer has to invoke the `__del__()` function which is
effectively a destructor from what I undestand. The fact that python uses a GC
makes timer object not to be deleted at the end of the scope it is initialized
in i.e. the end of the for loop. That results in awkward prints of elapsed time,
since the message is displayed upon `Timer`'s deletion.

This closes #10